requirements.txt
fastapi
pydantic
pydantic[email]
typing
SQLAlchemy
uvicorn
starlette
passlib
python-jose
python-dotenv
psycopg2-binary
streamlit